fix transition type bug and process bug

+++ utils/transition_sim.py
@@ -539,7 +539,7 @@ class NpyBuffer():
         # fig = visualizeTransition(agent, sim_obs, actions)
 
         is_expert = 1
-        transition = ExpertTransition(np.array(sim_states_new), sim_obs_new.astype(np.float32), sim_actions_new_star_idx[0].numpy(),
+        transition = ExpertTransition(np.array(sim_states_new), sim_obs_new.astype(np.float32), sim_actions_new_star_idx[0].numpy().astype(np.float32),
                                     sim_rewards2, sim_states2, sim_obs2, sim_dones2,
                                     sim_steps_lefts, np.array(is_expert))
         # if obs_type == 'pixel':
